// src/api/model/roleTypes.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Front interview quiz API
 * This is a simple app for preparation for frontend interview based on the OpenAPI 3.0 specification
 * OpenAPI spec version: 1.0.0
 */

export type RoleTypes = (typeof RoleTypes)[keyof typeof RoleTypes];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RoleTypes = {
  user: "user",
  teacher: "teacher",
  student: "student",
  admin: "admin",
} as const;